# docs/OPENAPI.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Dhruva API",
    "description": "Backend API for communicating with the Dhruva platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/services/inference/translation": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Translation",
        "operationId": "_run_inference_translation_services_inference_translation_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Serviceid",
              "type": "string"
            },
            "name": "serviceId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCATranslationInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCATranslationInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/transliteration": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Transliteration",
        "operationId": "_run_inference_transliteration_services_inference_transliteration_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Serviceid",
              "type": "string"
            },
            "name": "serviceId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCATransliterationInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCATransliterationInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/asr": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Asr",
        "operationId": "_run_inference_asr_services_inference_asr_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Serviceid",
              "type": "string"
            },
            "name": "serviceId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCAAsrInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCAAsrInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/tts": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Tts",
        "operationId": "_run_inference_tts_services_inference_tts_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Serviceid",
              "type": "string"
            },
            "name": "serviceId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCATtsInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCATtsInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/vad": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Vad",
        "operationId": "_run_inference_vad_services_inference_vad_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Serviceid",
              "type": "string"
            },
            "name": "serviceId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCAVadInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCAVadInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/ner": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Ner",
        "operationId": "_run_inference_ner_services_inference_ner_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "title": "Serviceid",
              "type": "string"
            },
            "name": "serviceId",
            "in": "query"
          },
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCANerInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCANerInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/s2s": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Sts",
        "operationId": "_run_inference_sts_services_inference_s2s_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCAS2SInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCAS2SInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/services/inference/pipeline": {
      "post": {
        "tags": [
          "Services"
        ],
        "summary": " Run Inference Pipeline",
        "operationId": "_run_inference_pipeline_services_inference_pipeline_post",
        "parameters": [
          {
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/TokenType"
                }
              ],
              "default": "API_KEY"
            },
            "name": "x-auth-source",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ULCAPipelineInferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ULCAPipelineInferenceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseErrorResponse": {
        "title": "BaseErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/exception__base_error___ErrorDetail"
          }
        }
      },
      "ClientErrorResponse": {
        "title": "ClientErrorResponse",
        "required": [
          "detail"
        ],
        "type": "object",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/exception__client_error___ErrorDetail"
          }
        }
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "TokenType": {
        "title": "TokenType",
        "enum": [
          "AUTH_TOKEN",
          "API_KEY"
        ],
        "description": "An enumeration."
      },
...
      "ULCATranslationInferenceRequest": {
        "title": "ULCATranslationInferenceRequest",
        "required": [
          "config",
          "input"
        ],
        "type": "object",
        "properties": {
          "controlConfig": {
            "title": "Controlconfig",
            "allOf": [
              {
                "$ref": "#/components/schemas/_ControlConfig"
              }
            ],
            "default": {
              "dataTracking": true
            }
          },
          "config": {
            "$ref": "#/components/schemas/_ULCATranslationInferenceConfig"
          },
          "input": {
            "title": "Input",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_ULCAText"
            }
          }
        }
      },
      "ULCATranslationInferenceResponse": {
        "title": "ULCATranslationInferenceResponse",
        "required": [
          "output"
        ],
        "type": "object",
        "properties": {
          "taskType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/_ULCATaskType"
              }
            ],
            "default": "translation"
          },
          "output": {
            "title": "Output",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_ULCATextPair"
            }
          },
          "config": {
            "$ref": "#/components/schemas/_ULCATranslationInferenceConfig"
          }
        }
      },
      "ULCATransliterationInferenceRequest": {
        "title": "ULCATransliterationInferenceRequest",
        "required": [
          "config",
          "input"
        ],
        "type": "object",
        "properties": {
          "controlConfig": {
            "title": "Controlconfig",
            "allOf": [
              {
                "$ref": "#/components/schemas/_ControlConfig"
              }
            ],
            "default": {
              "dataTracking": true
            }
          },
          "config": {
            "$ref": "#/components/schemas/_ULCATransliterationInferenceConfig"
          },
          "input": {
            "title": "Input",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/_ULCAText"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  }
}